+++ src/Distributed/distributed_architectures.jl
@@ -1,3 +1,5 @@
+using Oceananigans.Architectures
+
 using Oceananigans.Grids: validate_tupled_argument
 
 # TODO: Put connectivity inside architecture? MPI should be initialize so you can construct it in there.
